+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for labels and content type, [PR-50](https://github.com/reductstore/reduct-cpp/pull/50)
+
 ## [1.2.0] - 2022-12-20
 
 ### Added